+++ payment-service/src/main/java/n11bootcamp_project_backend/payment_service/enums/PaymentStatus.java
@@ -1,0 +1,7 @@
+package n11bootcamp_project_backend.payment_service.enums;
+
+public enum PaymentStatus {
+    PENDING,
+    SUCCESS,
+    FAILED
+}